Add recipe img new recipe structure and index link
Adds the img folder and a image for the first recipe used;
creates new recipe html with the basic html structure, an image of
the recipe and the structure the recipe-site and future recipes will
have in future commits.

Also add a link on index.html to the new recipe html file.